finmate-nextjs/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
google-generativeai==0.8.3
langfuse==2.58.3
feedparser==6.0.11
reportlab==4.2.5
PyPDF2==3.0.1
pydantic==2.10.3
python-multipart==0.0.20
openai==1.57.2
duckduckgo-search==5.3.0
yfinance==0.2.33
